refactor: shopping-cart 주소와 관련된 컴포넌트 이름 cartList로 수정
