+++ src/pages/CartList/PaymentAmount/style.ts
